# [Perf][MTP] Improve speculative decoding efficiency