--- a/custom_ops/gpu_ops/cpp_extensions.cc
+++ b/custom_ops/gpu_ops/cpp_extensions.cc
@@ -569,7 +569,7 @@ void SpeculateVerify(
     const paddle::Tensor &output_cum_offsets,
     const paddle::Tensor &actual_candidate_len,
     const paddle::Tensor &actual_draft_token_nums, const paddle::Tensor &topp,
-    int max_seq_len, int verify_window, bool enable_topp);
+    int max_seq_len, int verify_window, bool enable_topp, bool benchmark_mode);
 
 void SpeculateUpdateV3(const paddle::Tensor &seq_lens_encoder,
                        const paddle::Tensor &seq_lens_decoder,