**Reduced kernel launch latency:**

- The original custom op library was too large, causing kernel launch latency to reach ~200µs. By using PyBind instead, launch latency was reduced to ~20µs.

- Synchronization improvements:
Optimized several sync points in the speculative decoding path.  

**Conclusion**：
In the ERNIE-LITE + MTP pipeline, the time from verify to the next main model C-Embedding was reduced from 8.5ms to 4ms(113%).